camelCased the scripts, split them into seperate files, and other small refactors
