+++ vite/src/scripts/feed.js
@@ -1,0 +1,19 @@
+import { initializeApp } from "firebase/app";
+import { getAuth, createUserWithEmailAndPassword } from "firebase/auth";
+import { getDatabase, child, ref, set, get } from "firebase/database";
+import { getAnalytics } from "firebase/analytics";
+
+import firebaseConfig from "../../firebase.config.json";
+
+const app = initializeApp(firebaseConfig);
+const analytics = getAnalytics(app);
+const db = getDatabase();
+
+export const getFeedPosts = async () => {
+    const postsRef = ref(db, 'posts/')
+    try {
+        return (await get(postsRef)).val()
+    } catch {
+        return []
+    }
+}
